eduai-backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
pillow
httpx
sqlmodel
firebase-admin